requirements.txt
Flask_Login==0.2.11
Flask==0.11
Flask_SSLify==0.1.5
Flask_APScheduler==1.8.0
Flask_RESTful==0.3.6
Flask_Script==2.0.6
Werkzeug==0.14.1
APScheduler==3.5.1
python_dateutil==2.7.3
Appium-Python-Client==0.26
Jinja2==2.8
requests==2.18.4
robotframework==3.0.3
robotframework-appiumlibrary==1.4.6
robotframework-requests==0.4.7
robotframework-seleniumlibrary==3.1.1
robotframework-sshlibrary==3.1.1
robotframework-databaselibrary==1.0.1
Flask-Mail==0.9.0